meeting-scheduler-ai/requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.7.2
anthropic==0.8.1
PyYAML==6.0.1
python-dotenv==1.0.0
APScheduler==3.10.4
click==8.1.7
